Update progress log for home_screen refactor
